start moving timer typed into a central location
